a python script to sign pdfs
a python script to sign documents (signature, as in, handwritten signature). It uses the general convention that initials are signed on every page but full signature is only done once or twice in the document. You can also place a watermark or QR code on every day. You have to supply png files of each item.